Use '=' instead of '==' to test for equality, dash in gutsy (which seems
	* autogen.sh:
	  Use '=' instead of '==' to test for equality, dash in gutsy (which
	  seems to be the default shell now) doesn't seem to like '=='.
	* tests/conn_http_test.c:
	  Add cast to fix compiler warnings due to printf format.


svn path=/trunk/; revision=431

+++ tests/conn_http_test.c
@@ -410,11 +410,13 @@ test_simple_get (const gchar *url)
 	
 	if (ret == FALSE)
 	{
-		g_print("\t * gnet_http_get() failed (buflen = %u, code = %u).\n", buflen, code);
+		g_print("\t * gnet_http_get() failed (buflen = %u, code = %u).\n",
+		        (guint) buflen, code);
 		return FALSE;                
 	}
         
-	g_print("\t * gnet_http_get() succeeded (buflen = %u, code = %u).\n", buflen, code);
+	g_print("\t * gnet_http_get() succeeded (buflen = %u, code = %u).\n",
+	        (guint) buflen, code);
 
 	g_free (buf);
